Add initial Project IDX config
Project IDX (https://idx.dev/) is a way to easily create a development
environment for a coding project such as this.

I think it could be a good way to demo what we've built so far, and
perhaps be the default environment for adding/maintaining DMN rules in
the project.