Phl Property Tax Screener: hack to make benefits/checks more readable
I did this with a very complex expression (that Claude Sonnet 3.5 helped me create) which changes camelCase into "sentence" form (e.g. "Camel Case").

This is some ugly code added to the form field that displays the screening results; ideally we would find a way to encapsulate this functionality so the ugliness is contained to a function that can be called in multiple places.  I couldn't figure out how to do this for now, hence this solution.

Can be controlled by renaming the decision services in the DMN; will treat consecutive upper case letters as a single word/acronym.